First upload
The code is written to calculate the turbulent boundary layer trailing-edge(TE) noise, where the trailing edge can be either straight or serrated. The serration is described by the function x = hF(y/λ), where h is the half root to-tip amplitude and λ is the serration wavelength, and F(η) is the periodic shape function of the serration and has a period 1. This code only works for the sawtooth serration. The notations used in the code are consistent with the symbols used in the paper.